Layouts and home page in place
Even with only basic 960 formatting, it's beginning to take shape. Appropriate content is appearing where I expect it to.

The only problem is that on a post page, Liquid filters are being completely ignored. I wanted {{ page.date | date: "%Y-%m-%d" }}, but it
totally ignored the date formatting filter. Will look into it when it's
not 5 in the morning.