Added another post, tweaked home page date formatting
Nope, I'm just not happy with how this is turning out under Jekyll, and
here's the rant to prove it. Time to start over, though I'll be able to
use most of the layouts I've already developed for nyerm under the new
system with only a little modification. Well, that, and conversion to
HAML, which I much prefer working in, anyhow.