Converted a bunch of liquid on the homepage to erb
Starting to make some headway here. Figuring out what I've got exposed
is a bit of a trick after months away from this code, but it's all
coming back to me. I want a bunch of convenience methods attached to the
Site, Page, and Post classes, and there are some things that don't seem
to work as I'd expect.

In particular, post.content is spitting back the post after processing
through its entire layout, which isn't quite what I want. Some
processing is necessary, of course, like converting Markdown to HTML,
but the header and footer shouldn't show up when embedding that content.
Perhaps I just don't have the right header information in the layouts
yet.

Also, above_fold needs to return the entire body of the post if there
isn't a more marker. Currently, both above_fold and below_fold are
empty if there's no marker, and they should contain something. That, or
the template code should check to see if above_fold is nil and print
the full contents if that's the case.